REFACTOR: Breaking down the UI into reusable components
