+++ src/scss/assignment.scss
@@ -1,0 +1,208 @@
+.assignment-component {
+  background-color: #1f2666;
+  border-radius: 8px;
+  width: 100%;
+  height: 719px;
+  box-shadow: 0 3px 25px #0000000f;
+
+  padding: 37px 0 17px;
+  padding-right: 3px;
+
+  h3 {
+    color: #ffffff;
+    font-size: 18px;
+    font-weight: 600;
+    margin-bottom: 42px;
+    padding: 0 31px;
+  }
+
+  .article-details {
+    .detail-item {
+      padding: 0 31px;
+      margin-bottom: 43px;
+
+      & > div:nth-child(1) {
+        width: 25%;
+
+        svg {
+          width: 24px;
+          height: 24px;
+          color: #0266f4;
+          cursor: pointer;
+        }
+
+        p {
+          margin: 0;
+          font-size: 12px;
+          color: #ffffff;
+        }
+      }
+      & > div:nth-child(2) {
+        width: 21px;
+        height: 21px;
+        cursor: pointer;
+
+        img {
+          width: 100%;
+          height: 100%;
+          object-fit: contain;
+        }
+      }
+    }
+
+    .detail-schedule {
+      padding: 0 31px;
+      height: 526px;
+      overflow-y: scroll;
+      &::-webkit-scrollbar {
+        width: 3px;
+      }
+
+      &::-webkit-scrollbar-track {
+        background: #cccccc;
+      }
+
+      & > section {
+        margin-bottom: 40px;
+        p {
+          color: #ffffff;
+          font-weight: 600;
+          font-size: 12px;
+          margin-bottom: 20px;
+        }
+
+        article {
+          & > div {
+            background-color: #ffffff;
+            border: 0.5px solid #cccccc;
+            border-radius: 4px;
+            padding: 14px 12px;
+            margin-bottom: 20px;
+            align-items: center;
+            gap: 29px;
+
+            & > div:nth-child(1) {
+              width: 36px;
+              height: 36px;
+              padding: 11px;
+              background: #1f266633;
+              border-radius: 4px;
+
+              img {
+                width: 100%;
+                height: 100%;
+                object-fit: contain;
+              }
+            }
+
+            & > div:nth-child(2) {
+              p {
+                font-size: 12px;
+                margin: 0;
+                color: #748aa1;
+              }
+
+              h6 {
+                font-size: 12px;
+                color: #0266f4;
+                margin-bottom: 8px;
+              }
+            }
+          }
+        }
+      }
+    }
+  }
+}
+
+.assignment-modalContainer {
+  position: fixed;
+  right: 0;
+  bottom: 0;
+  top: 0;
+  width: 100%;
+  height: 100%;
+  background: rgba(0, 0, 0, 0.7);
+  margin: 0;
+  overflow-y: hidden;
+}
+
+@media screen and (max-width: 1200px) {
+  .assignment-component {
+    .article-details {
+      .detail-item {
+        & > div:nth-child(1) {
+          width: 35%;
+        }
+      }
+    }
+  }
+}
+
+@media screen and (max-width: 998px) {
+  .assignment-modalContainer {
+    .assignment-modal {
+      width: 100%;
+      padding-top: 70px;
+
+      & > article {
+        article {
+          & > div {
+            label {
+              justify-content: flex-start;
+            }
+          }
+        }
+      }
+    }
+  }
+
+  .assignment-component {
+    width: 100%;
+    margin-bottom: 20px;
+
+    .article-details {
+      .detail-item {
+        & > div:nth-child(1) {
+          width: 20%;
+        }
+      }
+    }
+  }
+}
+
+@media screen and (max-width: 550px) {
+
+    .assignment-component {
+
+        .article-details {
+          .detail-item {
+            & > div:nth-child(1) {
+              width: 30%;
+            }
+          }
+        }
+    }
+
+  .assignment-modalContainer {
+    .assignment-modal {
+      padding: 34px 10px;
+
+      & > article {
+        & > div:nth-child(1) {
+          h3 {
+            margin-left: 40px;
+          }
+        }
+
+        article {
+          & > div {
+            label {
+              padding-left: 10px;
+            }
+          }
+        }
+      }
+    }
+  }
+}
